requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.11.0
python-dotenv==1.0.0
pydantic>=2.5.0
psycopg2-binary==2.9.7
pgvector==0.2.4
python-multipart==0.0.6
litellm==1.74.3
pydantic-settings==2.1.0
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
chardet==5.2.0
anyio==3.7.1